requirements.txt
# ============================================================================
# Texlink Analytics Platform — Python Dependencies
# Python 3.11+
# ============================================================================

# --- Database ---
psycopg2-binary==2.9.9
sqlalchemy==2.0.36

# --- Data Generation ---
faker==26.0.0
pyyaml==6.0.2

# --- Data Processing ---
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1

# --- Statistical / ML ---
pymc==5.17.0
scikit-learn==1.5.2
lifetimes==0.11.3         # BG/NBD model for CLV

# --- Dashboard ---
streamlit==1.39.0
plotly==5.24.1
altair==5.4.1
pydeck==0.9.1             # Maps for geographic analysis

# --- Utilities ---
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.67.0
rich==13.9.4              # Pretty terminal output

# --- Dev / Testing ---
pytest==8.3.3
pytest-cov==6.0.0
black==24.10.0
ruff==0.7.4
